requirements.txt
Django==3.2.16
djangorestframework==3.14.0
django-cors-headers==3.4.0
psycopg2==2.8.6
gunicorn==20.0.4
keyboa==3.1.0
python-telegram-bot-calendar
pyTelegramBotAPI
drf-spectacular
Pillow
djoser
djangorestframework_simplejwt
django-filter